Expanded x86 code generation
- finished target-independent code generation
- finished x86 code generation, except calling conventions:
  - cdecl (32-bit, Linux)
  	- function calls: done
    - function entry/exit: work in progress: NIY
    - builtin functions for variadic arguments: NIY
  - System V ABI (64-bit, Linux): NIY
  - Microsoft x64 (64-bit, Windows): NIY
- added global object serialization
- added stack allocation
- added prologue/epilogue insertion
- added __builtin_va_list (alias for void*)
- fixed goto/jump-to-jump optimization
- fixed logical AND and logical OR
- fixed IRs in CFG without a result
- fixed return value in CFG

+++ src/main/java/at/syntaxerror/syntaxc/generator/arch/Alignment.java
@@ -35,11 +35,11 @@ public abstract class Alignment {
 	
 	public abstract StructAlignment getMemberAlignment(StructType struct, Type type, int offset, int bitOffset, int bitWidth, boolean wasBitfield);
 	
-	public static int alignAt(int offset, int alignment) {
+	public static long alignAt(long offset, long alignment) {
 		if(alignment < 2)
 			return offset;
 		
-		int pad = offset % alignment;
+		long pad = offset % alignment;
 		
 		return pad == 0
 			? offset
